magento/inventory#2839: Moved to filters extensions attributes
