+++ InventoryInStorePickup/etc/extension_attributes.xml
@@ -6,7 +6,7 @@
  */
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Api/etc/extension_attributes.xsd">
-    <extension_attributes for="Magento\InventoryInStorePickupApi\Api\Data\SearchRequestInterface">
-        <attribute code="products_info" type="Magento\InventoryInStorePickupShippingApi\Api\Data\ProductInfoInterface[]"/>
+    <extension_attributes for="Magento\InventoryInStorePickupApi\Api\Data\SearchRequest\FiltersInterface">
+        <attribute code="products_info" type="Magento\InventoryInStorePickupApi\Api\Data\SearchRequest\ProductInfoInterface[]"/>
     </extension_attributes>
 </config>
