+++ InventoryInStorePickupFrontend/view/frontend/web/js/view/store-selector.js
@@ -163,7 +163,11 @@ define([
 
             _.each(items, function (item) {
                 if (item['qty_options'] === undefined || item['qty_options'].length === 0) {
-                    productsInfo.push({sku: item.sku});
+                    productsInfo.push(
+                        {
+                            sku: item.sku
+                        }
+                    );
                 }
             });
 
@@ -173,8 +177,10 @@ define([
                         radius: this.nearbySearchRadius,
                         searchTerm: searchQuery
                     },
-                    extension_attributes: {
-                        productsInfo: productsInfo
+                    filters: {
+                        extension_attributes: {
+                            productsInfo: productsInfo
+                        }
                     },
                     pageSize: this.nearbySearchLimit
                 })
