+++ InventoryInStorePickupShipping/composer.json
@@ -7,8 +7,6 @@
         "magento/module-inventory-in-store-pickup-shipping-api": "*",
         "magento/module-inventory-in-store-pickup-api": "*",
         "magento/module-inventory-sales-api": "*",
-        "magento/module-inventory-api": "*",
-        "magento/module-inventory": "*",
         "magento/module-quote": "*",
         "magento/module-store": "*",
         "magento/module-catalog": "*"
